requirements.txt
# Basic FastAPI requirements
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.10.7
python-multipart==0.0.6

# Google API requirements
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.0.2

# Additional utilities
requests==2.28.2
python-jose==3.3.0